Use go1.6 template inheritance.
This moves all of the <html></html> out of header and footer into a new
base template, which is now the main template loaded on controllers.

Content can be added into the base template using {{block "content" .}}.

Header and footer are still here for now, but have been reduced to their
non-HTMLy content: footer is currently empty.