Make RenderTemplate pass errors on.
This is slightly less clean, but means errors all arrive at the
same place for now.

Presumably, at some point, we'll want errors to trigger a new
template render for an error page.  It might be useful to make this
a new wrapper function over current RenderTemplate rather than to
shoehorn this into RenderTemplate itself...?