PURGE ALL THE NODE
OK, so we've lost a couple of things here, like automatic minifying of
the JS and automatic recompilation of the scss, but I believe getting
rid of the largely unnecessary use of Node makes it a net positive.

I've moved the scss to a "sass/" folder since that's now the only thing
that's being generated, and set the templates so they're expecting a
*.scss.css file, mostly just so we can effectively git ignore them.

I'm not hugely familiar with said libraries, so let me know if I've
missed anything obvious.