Keep entire day range in WeekSchedule
WeekSchedule now tracks the entire set of days it manages, as
opposed to just the start and end day.  It also uses URY start
times in more places, as opposed to local midnight.

This simplifies a bit of the logic, as well as making us able to
reduce the amount of hardcoding in the template.

Technically, this means we could generate 9-day schedules (or
n-day schedules for arbitrary n), but whether this is a good thing
is debatable.