Strip HTML in tooltip descriptions
MyRadio sends show descriptions as full-on HTML5, which is great
for show description pages, but bad for tooltips.  This commit
teaches 2016-site to strip the HTML when emitting descriptions in
tooltips.

This is a fairly heavyweight way to do this, involving manually
invoking a HTML parser, but it seems to work ok.