Use router reversing for timeslot URLs
This means we don't need to construct the URLs ourselves, which
saves repetition and a source of errors later on, but _does_ mean
we have to pass the router around into places from which it used to
be separate.

This is a bit more invasive than I'd like, and could probably be
refactored a lot.